tg_bot/requirements.txt
aiogram==3.0.0rc5
aioredis==2.0.1
python-dotenv==1.0.0
